run process.php and gather the shasum

+++ ansible/octo-weasel/run-performance-test.sh
@@ -2,6 +2,8 @@
 
 echo "$(date '+%Y-%m-%d %H-%M-%S') Starting ..."
 
+currentTime=$(date +%Y-%m-%d.%H-%M-%S)
+shaSum=""
 
 function execute_tests {
 
@@ -21,20 +23,26 @@ function execute_tests {
   cd /var/www/owncloud
   git fetch
   git checkout -q $2 || exit 1
+  shaSum=$(git rev-parse HEAD)
   git submodule update
 
   echo "$(date '+%Y-%m-%d %H-%M-%S') Install owncloud ..."
   sudo -u www-data php occ maintenance:install --admin-pass=admin --database=mysql --database-name=owncloud --database-user=owncloud --database-pass=owncloud
 
   mkdir -p /tmp/performance-tests
-  currentTime=$(date +%Y-%m-%d.%H-%M-%S)
   echo "$(date '+%Y-%m-%d %H-%M-%S') Running performance test ..."
   DAV_USER=admin DAV_PASS=admin /root/administration/performance-tests-c++/webdav-benchmark http://localhost/remote.php/webdav/ -csv > /tmp/performance-tests/$currentTime.csv
 
-  if [ -f /var/log/mysql/mysql.log ]; then
+  if [ "$1" -eq "1" ]; then
+    rm /tmp/performance-tests/$currentTime.csv
+  else
+    mv /tmp/performance-tests/$currentTime.csv /tmp/performance-tests/result.$currentTime.csv
+  fi
+
+  if [ -f /var/log/mysql/mysql.log -a "$1" -eq "1" ]; then
     mv /var/log/mysql/mysql.log /tmp/performance-tests/mysql-general-query-$currentTime.log
   fi
-  if [ -f /var/log/mysql/mysql-slow.log ]; then
+  if [ -f /var/log/mysql/mysql-slow.log -a "$1" -eq "0" ]; then
     mv /var/log/mysql/mysql-slow.log /tmp/performance-tests/mysql-slow-query-$currentTime.log
   fi
 
@@ -50,5 +58,13 @@ echo "$(date '+%Y-%m-%d %H-%M-%S') Running WITHOUT general query logger ... "
 execute_tests 0 $1
 echo "$(date '+%Y-%m-%d %H-%M-%S') Running WITH general query logger ... "
 execute_tests 1 $1
+
+echo "$(date '+%Y-%m-%d %H-%M-%S') Aggregate query and access logs ..."
+cd /root
+php process.php /tmp/performance-tests/mysql-general-query-$currentTime.log /tmp/performance-tests/access-$currentTime.log /tmp/performance-tests/results-$currentTime.json
+
+# TODO push the results to weasel API
+
+echo "SHA sum: $shaSum"
 
 echo "$(date '+%Y-%m-%d %H-%M-%S') Finished"
